leaks

+++ utils/ft_substr.c
@@ -14,18 +14,18 @@
 
 char	*ft_substr(char *s, int start, int len)
 {
-	char			*subs;
-	int	i;
-	int				length;
+	char	*subs;
+	int		i;
+	int		length;
 
 	i = 0;
 	if (s == NULL)
 		return (NULL);
-	length = ft_strlen(s) - (int)start;
+	length = ft_strlen(s) - start;
 	if (length <= 0)
 		length = 0;
-	else if (length > (int)len)
-		length = (int)len;
+	else if (length > len)
+		length = len;
 	subs = (char *)malloc(sizeof(char) * (length + 1));
 	if (subs == NULL)
 		return (NULL);
